[Receipt Bin] Fix `per` param
We use the `per` param to determine the layout of the page (whether transactions are rendered in groups, or a single list). Unfortunately, it 500s at the moment because we're comparing an integer to a string (`params[:per` is a string).

This fixes it by casting `per` to an int.