[Admin] Add tags when creating organizations (#10850)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->
Currently you can not tag organizations when creating them and rather
have to go into the admin tab of the organization and create it there.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
created a drop down multiselect that allows you to select multiple
organization tags when creating an organization. Removed the 2 buttons
that says created by teenagers and hack clubbers since it's now in the
multiselect.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

![image](https://github.com/user-attachments/assets/2d9d4804-4d2a-4bb7-81fa-76882a16cfbc)