Automate airtable application status (#10827)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. --> Onboarders have to
manually set the status of an application to "documents sent" and
"onboarded" and it's a little annoying to do that knowing it could be
automated.



## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. --> I automated it. "Documents sent" after a
contract signee is invited and "Onboarded" after the org is activated.



<!-- If there are any visual changes, please attach images, videos, or
gifs. -->