Fix dark mode blog widget error (#10998)
## Summary of the problem
<!-- Why these changes are being made? What problem does it solve? Link
any related issues to provide more details. -->

https://appsignal.com/hack-club/sites/6596247683eb67648f30f807/exceptions/incidents/2013/samples/timestamp/2025-07-16T18:55:27Z

```
Error: Missing outlet element "blog" for host controller "dark-mode-toggle". Stimulus couldn't find a matching outlet element using selector "#blog-widget".
```

This was caused by places where the theme toggle is rendered without the
blog widget

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Added checks to make sure the dark mode toggle controller doesn't try to
update the blog widget when it is not present. Also removed the Flipper
flag on the blog widget and removed the need to be signed in to see the
blog widget.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->