Add `category_slug` and `category_label` to Event's transaction CSV export.
Note: In `TransactionCategory`, I'm delegating the `label` method to `definition`.

```ruby
delegate :label, to: :definition

def definition
  TransactionCategory::Definition::ALL[slug]
end
```

If for some reason the slug can't be looked up and `definition` is `nil`, then we will get a NoMethodError when calling `label` on `nil`.

I'm not sure how likely it is for `definition` to return `nil`. If it's possible, then we should explicitly handle that.