[v4 API][Check Deposit] Check policy before rendering check images
On the normal web UI, we only show the check deposit image to managers or auditors. This means that members and readers aren't able to view it.

On the v4 API, we weren't performing the same policy check. As a result, all organizers of the event could view check deposit images.

This PR fixes the bug by checking `CheckDepositPolicy#view_image?` before rendering the check deposit image URLs.