[Check Deposit] Allow depositor to view check images even if they're a member/reader.
Generally speaking, we hide check deposit images from readers and members because it contains the account & routing number of the donor (check sender).

However, we also allow members to upload check deposits. With our previous pundit policy, members would upload check deposits, but then wouldn't be able to see the image they uploaded.

This PR fixes that by allowing a user to view the check deposit images as long as they're an organizer (e.g. reader, member, etc.) AND also was the person who uploaded the check deposit.

This means the final pundit policy looks like:
```ruby
def view_image?
  # You can view the check deposit images (front & back) as long as you meet
  # at least one of the following conditions:
  # - You're an auditor (admin)
  # - You're a manager of the event
  # - You're an organizer of the event (e.g. reader, member, etc.), but ALSO
  #   was the person who uploaded the check deposit.
  auditor_or_manager? || (user? && record.created_by == user)
end
```

N.B. The "you're an organizer of the event" check is important because it prevents a user who uploaded a check deposit, but is later kicked from the organization, from accessing the check deposit images.

Although rare, this logic also intentionally allows for the following:
1. Member uploads a check deposit => they can view the check deposit images they uploaded.
2. The member gets demoted to reader => they can still view the check deposit images they uploaded.
3. The reader is removed from the org => they **lose access** to the images.