Fix errors found by raise_on_missing_callback_actions=true
The index action could not be found for the :set_ach_transfer callback on AchTransfersController, but it is listed in the controller's :except option.
The new action could not be found for the :set_event callback on EventsController, but it is listed in the controller's :except option.
The create action could not be found for the :set_event callback on EventsController, but it is listed in the controller's :except option.