Made acronym expansions show to all teenagers
The acronym expansions only showed to users defined as hack_clubber, but not everyone is tagged as a Hack Clubber in the HCB backend (seemingly only tagged if you are a member of an HCB org tagged as organized by Hack Clubbers), so not every Hack Clubber can see the expansions. I updated it to show if they're a Hack Clubber or if they're a teenager, so that it shows to everyone except for adults who might find it weird.