[HQ] Temporarily disable filtering of HQ ledger in transparency mode
We are seeing 16x more traffic due to people web scraping/crawling the HQ ledger. Unfortunately, this is so much more traffic than we're used to and is causing performance issues for normal users.

Logs show them crawling every filter variation on the ledger. To counteract this, I'm temporarily disabling all filters (except search) for the HQ ledger in transparency mode. It's important to note that **all transactions are still accessible in transparency mode**. If you want to filter, I'd suggest just downloading a CSV and filtering in Excel.